+++ business/src/main/java/com/yun/train/resp/DailyTrainTicketQueryResp.java
@@ -75,7 +75,7 @@ public class DailyTrainTicketQueryResp {
     /**
      * 一等座票价
      */
-    private Integer ydzPrice;
+    private BigDecimal ydzPrice;
 
     /**
      * 二等座余票
@@ -85,7 +85,7 @@ public class DailyTrainTicketQueryResp {
     /**
      * 二等座票价
      */
-    private Integer edzPrice;
+    private BigDecimal edzPrice;
 
     /**
      * 软卧余票
@@ -215,11 +215,11 @@ public class DailyTrainTicketQueryResp {
         this.ydz = ydz;
     }
 
-    public Integer getYdzPrice() {
+    public BigDecimal getYdzPrice() {
         return ydzPrice;
     }
 
-    public void setYdzPrice(Integer ydzPrice) {
+    public void setYdzPrice(BigDecimal ydzPrice) {
         this.ydzPrice = ydzPrice;
     }
 
@@ -231,11 +231,11 @@ public class DailyTrainTicketQueryResp {
         this.edz = edz;
     }
 
-    public Integer getEdzPrice() {
+    public BigDecimal getEdzPrice() {
         return edzPrice;
     }
 
-    public void setEdzPrice(Integer edzPrice) {
+    public void setEdzPrice(BigDecimal edzPrice) {
         this.edzPrice = edzPrice;
     }
